Added check for Java9 to Gradle wrapper
I want to use the same ./gradlew for both my system and our Travis
builds.  But it's a pain to install Java9 in Travis and we don't really
need it anyway for a JAR, so this updated script will check and add the
"--illegal-access=permit" option only for Java9.

Also added an output message for whether the files uploaded to bintray
will be published as well.  (I just wanted confirmation in the output
for logging purposes.)