Added Travis build script
The script is needed to test TRAVIS_TAG and build appropriately.  If
there is no tag, the build is just to confirm that there aren't any
build issues.  If there *IS* a tag, then we also want to upload to
bintray and pubish the results, using the tag string as the version
string to publish it under.  This can be different than the build
version.  So, "1.4.1.9" might be the development build version, but we
could choose to publish it as "1.4-dev-latest".