Moved bintray deployment from Gradle to Travis
This involved removing the configuration from build.gradle and
adding a 'deploy' block to .travis.yml.  Also, a deployment descriptor
for bintray was added to packages/template.deploy.json with a
corresponding task in Gradle to substitute repo- and version-specific
information into the file for publishing.

The bintray user & key must be specified using Travis settings for the
repo, along with the deployment key for GitHub releases as well.